some betters

+++ KCalMeter/IKCalMeterService.cs
@@ -2,7 +2,7 @@
 
 public interface IKCalMeterService
 {
-    Task WritePortion(string name, int portion, long chatId, long userId, DateTime messageDate);
+    Task WritePortion(string name, decimal portion, long chatId, long userId, DateTime messageDate);
     Task DeleteFood(string name, long chatId, long userId);
     Task WriteFood(string name, int basePortion, FoodInfo foodInfo, long chatId, long userId);
     Task<FoodMessageDbo[]> SelectPortions(int count, long chatId, long userId);
